refactor cloudflare_configuration.go
