+++ internal/clients/cf/configuration.go
@@ -1,4 +1,4 @@
-package controller
+package cf
 
 import (
 	"time"
